Rename `ClientTypeHelper` as `ClientTypeExtensions`
